tools-extra: Added FAT16 SD generator
Why do I do this to myself.
I could have made a lot of nice things.
One project I have is to make a telegram bot that checks for deals on
games. It could help me and many others save money by catching sales and
allowing them to play games they really want.

But no, my hobby is dealing with FAT16 in the year 2018.
Is this why no one wants to be with me?